+++ src/languages/javascript.js
@@ -5,6 +5,8 @@ Category: common, scripting
 Website: https://developer.mozilla.org/en-US/docs/Web/JavaScript
 */
 
+import * as ECMAScript from "./lib/ecmascript";
+
 export default function(hljs) {
   var FRAGMENT = {
     begin: '<>',
@@ -16,24 +18,9 @@ export default function(hljs) {
   };
   var IDENT_RE = '[A-Za-z$_][0-9A-Za-z$_]*';
   var KEYWORDS = {
-    keyword:
-      'in of if for while finally var new function do return void else break catch ' +
-      'instanceof with throw case default try this switch continue typeof delete ' +
-      'let yield const export super debugger as async await static ' +
-      // ECMAScript 6 modules import
-      'import from as'
-    ,
-    literal:
-      'true false null undefined NaN Infinity',
-    built_in:
-      'eval isFinite isNaN parseFloat parseInt decodeURI decodeURIComponent ' +
-      'encodeURI encodeURIComponent escape unescape Object Function Boolean Error ' +
-      'EvalError InternalError RangeError ReferenceError StopIteration SyntaxError ' +
-      'TypeError URIError Number Math Date String RegExp Array Float32Array ' +
-      'Float64Array Int16Array Int32Array Int8Array Uint16Array Uint32Array ' +
-      'Uint8Array Uint8ClampedArray ArrayBuffer DataView JSON Intl arguments require ' +
-      'module console window document Symbol Set Map WeakSet WeakMap Proxy Reflect ' +
-      'Promise setInterval setTimeout clearInterval clearTimeout'
+    keyword: ECMAScript.KEYWORDS.join(" "),
+    literal: ECMAScript.LITERALS.join(" "),
+    built_in: ECMAScript.BUILT_INS.join(" ")
   };
   var NUMBER = {
     className: 'number',